+++ components/global/InfoBox.vue
@@ -1,9 +1,7 @@
 <template>
   <div class="bg-primary-500 text-base-content p-4 mb-4">
     <p>
-      <slot name="info-box">
-        default
-      </slot>
+      <slot name="info-box"> default </slot>
     </p>
   </div>
 </template>