+++ components/global/PrevNext.vue
@@ -1,18 +1,10 @@
 <template>
   <div class="flex justify-between gap-10">
-    <NuxtLink
-      v-if="next"
-      :to="next.slug"
-      class="font-bold hover:underline"
-    >
+    <NuxtLink v-if="next" :to="next.slug" class="font-bold hover:underline">
       ← {{ next.title }}
     </NuxtLink>
     <span v-else>&nbsp;</span>
-    <NuxtLink
-      v-if="prev"
-      :to="prev.slug"
-      class="font-bold hover:underline"
-    >
+    <NuxtLink v-if="prev" :to="prev.slug" class="font-bold hover:underline">
       {{ prev.title }} →
     </NuxtLink>