+++ plugins/markdownit.js
@@ -4,7 +4,6 @@
 
 import { defineNuxtPlugin } from '#app'
 
-
 import hljs from 'highlight.js'
 import mdit from 'markdown-it'
 
@@ -16,25 +15,30 @@ import def from 'markdown-it-deflist'
 import ins from 'markdown-it-ins'
 import container from 'markdown-it-container'
 
-
 const markdownit = new mdit({
-  html:         true,
-  xhtmlOut:     false,
-  breaks:       false,
-  langPrefix:   'language-',
-  linkify:      true,
-  typographer:  true,
+  html: true,
+  xhtmlOut: false,
+  breaks: false,
+  langPrefix: 'language-',
+  linkify: true,
+  typographer: true,
   quotes: '“”‘’',
   highlight: function (str, lang) {
     if (lang && hljs.getLanguage(lang)) {
       try {
-        return '<pre class="hljs"><code>' +
+        return (
+          '<pre class="hljs"><code>' +
           hljs.highlight(str, { language: lang, ignoreIllegals: true }).value +
           '</code></pre>'
-      } catch (error) {console.log(error)}
+        )
+      } catch (error) {
+        console.log(error)
+      }
     }
 
-    return '<pre class="hljs"><code>' + mdit.utils.escapeHtml(str) + '</code></pre>'
+    return (
+      '<pre class="hljs"><code>' + mdit.utils.escapeHtml(str) + '</code></pre>'
+    )
   }
 })
   .use(sub)
@@ -43,11 +47,10 @@ const markdownit = new mdit({
   .use(emo)
   .use(def)
   .use(ins)
-  .use(container,'codeblock',{marker:'@'})
+  .use(container, 'codeblock', { marker: '@' })
 
 markdownit.linkify.set({ fuzzyEmail: false })
 
-
-export default defineNuxtPlugin(nuxtApp => {
-  nuxtApp.provide('mdit',markdownit)
+export default defineNuxtPlugin((nuxtApp) => {
+  nuxtApp.provide('mdit', markdownit)
 })