+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog Top Nuxt3, by @ctwhome
 
+### Jan 28, 2022
+
+- 🎉 Add prettier and improve eslint config
+
 ### Jan 8, 2022
+
 - 🎉 Add styled Markdown files